Enable configuring HTTP transport retry behavior and add `request_timeout` parameter to `get_detector` (#386)
This PR includes two distinct changes.
- Adds a new parameter `http_transport_retries` to the `Groundlight`
constructor, enabling the retry behavior to be modified when
initializing a `Groundlight` instance.
- The retry behavior specified will apply to all API requests made
through the `Groundlight` instance.
- Currently no way is provided to configure the retry behavior
per-request, or to change it once the `Groundlight` instance has been
initialized (this is technically possible but awkward).
- I added a test to verify that we can pass the parameter in and it gets
passed through to the inner API client.
- Adds a `request_timeout` parameter to the `get_detector` method which
enables changing the request timeout values for the request.
- An identical parameter was previously added for `submit_image_query`.
Eventually I think we'll want to make this available for all methods,
but for now we don't need that.
  - I added a test to verify that this parameter works properly.

These changes are intended to be utilized by the edge endpoint to
improve behavior under poor network connectivity.