Add bounding boxes label enum (#390)
This PR adds a new bounding boxes label enum to the SDK. It will be used
to correctly construct responses on the edge, like `Label` is used for
binary mode. I also bumped the version number to do a new release with
this enum.

Due to compatibility issues between `datamodel-code-generator` and the
version of pydantic in use, I also update `datamodel-code-generator`.